+++ src/main/java/com/akkin/gulbi/delete/presentation/GulbiDeleteApiControllerDocs.java
@@ -1,19 +1,18 @@
-package com.akkin.main;
+package com.akkin.gulbi.delete.presentation;
 
-import com.akkin.main.dto.MainResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
-import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import javax.servlet.http.HttpServletRequest;
+import org.springframework.http.ResponseEntity;
 
-@Tag(name = "메인 정보", description = "메인 정보 API")
-public interface MainControllerDocs {
+@Tag(name = "아낀 항목 삭제", description = "아낀 항목 삭제 API")
+public interface GulbiDeleteApiControllerDocs {
 
-    @Operation(summary = "메인 정보", parameters = {
+    @Operation(summary = "아낀 항목 삭제", parameters = {
         @Parameter(
             in = ParameterIn.HEADER,
             name = "accessToken",
@@ -25,13 +24,14 @@ public interface MainControllerDocs {
             name = "refreshToken",
             required = true,
             schema = @Schema(type = "string"),
-            description = "Refresh Token")
+            description = "Refresh Token"),
+        @Parameter(
+            in = ParameterIn.PATH,
+            name = "id",
+            required = true,
+            schema = @Schema(type = "long"),
+            description = "아낀 항목 식별자")
     },
-        responses = {
-            @ApiResponse(responseCode = "200", description = "메인 정보에 쓸 데이터",
-                content = {@Content(mediaType = "application/json",
-                    schema = @Schema(implementation = MainResponse.class))
-                })
-        })
-    MainResponse enterMain(HttpServletRequest request);
+        responses = {@ApiResponse(responseCode = "200", description = "성공적으로 아낀 항목을 삭제하였습니다.")})
+    ResponseEntity<Void> deleteGulbi(Long gulbiId, HttpServletRequest request);
 }
