+++ src/main/java/com/akkin/gulbi/weekly/presentation/WeeklyApiController.java
@@ -1,7 +1,8 @@
-package com.akkin.gulbi.monthly;
+package com.akkin.gulbi.weekly.presentation;
 
 import com.akkin.auth.aop.AuthRequired;
-import com.akkin.gulbi.monthly.dto.MonthlyResponse;
+import com.akkin.gulbi.weekly.WeeklyService;
+import com.akkin.gulbi.weekly.dto.MemberWeeklyResponse;
 import com.akkin.auth.dto.AuthMember;
 import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
@@ -12,14 +13,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @RequestMapping("/api/members")
 @RestController
-public class MonthlyController implements MonthlyControllerDocs {
+public class WeeklyApiController implements WeeklyApiControllerDocs {
 
-    private final MonthlyService monthlyService;
+    private final WeeklyService weeklyService;
 
+    @Override
     @AuthRequired
-    @GetMapping("/monthly")
-    public MonthlyResponse getMonthInfo(HttpServletRequest request) {
+    @GetMapping("/weekly")
+    public MemberWeeklyResponse getGulbis(HttpServletRequest request) {
         AuthMember authMember = (AuthMember) request.getAttribute("authMember");
-        return monthlyService.getMonthInfo(authMember.getId());
+        return weeklyService.getWeekly(authMember.getId());
     }
 }
