+++ src/main/java/com/akkin/main/presentation/MainApiControllerDocs.java
@@ -1,18 +1,19 @@
-package com.akkin.gulbi.delete;
+package com.akkin.main.presentation;
 
+import com.akkin.main.dto.MainResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
+import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import javax.servlet.http.HttpServletRequest;
-import org.springframework.http.ResponseEntity;
 
-@Tag(name = "아낀 항목 삭제", description = "아낀 항목 삭제 API")
-public interface GulbiDeleteControllerDocs {
+@Tag(name = "메인 정보", description = "메인 정보 API")
+public interface MainApiControllerDocs {
 
-    @Operation(summary = "아낀 항목 삭제", parameters = {
+    @Operation(summary = "메인 정보", parameters = {
         @Parameter(
             in = ParameterIn.HEADER,
             name = "accessToken",
@@ -24,14 +25,13 @@ public interface GulbiDeleteControllerDocs {
             name = "refreshToken",
             required = true,
             schema = @Schema(type = "string"),
-            description = "Refresh Token"),
-        @Parameter(
-            in = ParameterIn.PATH,
-            name = "id",
-            required = true,
-            schema = @Schema(type = "long"),
-            description = "아낀 항목 식별자")
+            description = "Refresh Token")
     },
-        responses = {@ApiResponse(responseCode = "200", description = "성공적으로 아낀 항목을 삭제하였습니다.")})
-    ResponseEntity<Void> deleteGulbi(Long gulbiId, HttpServletRequest request);
+        responses = {
+            @ApiResponse(responseCode = "200", description = "메인 정보에 쓸 데이터",
+                content = {@Content(mediaType = "application/json",
+                    schema = @Schema(implementation = MainResponse.class))
+                })
+        })
+    MainResponse enterMain(HttpServletRequest request);
 }
